feat(htmlToSlate): extract config
