+++ src/config/slatetoDom/default.ts
