+++ src/serializers/htmlToSlate/index.ts
@@ -5,47 +5,11 @@ import { getAttributeValue, getChildren, getName, textContent } from 'domutils'
 import { hasLineBreak } from '../../utilities'
 import { IattributeMap } from '../../types'
 
-interface ItagMap {
-  [key: string]: (a?: Element) => object
-}
-
-const ELEMENT_TAGS: ItagMap = {
-  a: (el) => ({
-    type: 'link',
-    newTab: el && getAttributeValue(el, 'target') === '_blank',
-    url: el && getAttributeValue(el, 'href'),
-  }),
-  blockquote: () => ({ type: 'blockquote' }),
-  h1: () => ({ type: 'h1' }),
-  h2: () => ({ type: 'h2' }),
-  h3: () => ({ type: 'h3' }),
-  h4: () => ({ type: 'h4' }),
-  h5: () => ({ type: 'h5' }),
-  h6: () => ({ type: 'h6' }),
-  li: () => ({ type: 'li' }),
-  ol: () => ({ type: 'ol' }),
-  p: () => ({ type: 'p' }),
-  ul: () => ({ type: 'ul' }),
-}
-
-const TEXT_TAGS: ItagMap = {
-  code: () => ({ code: true }),
-  pre: () => ({ code: true }),
-  del: () => ({ strikethrough: true }),
-  em: () => ({ italic: true }),
-  i: () => ({ italic: true }),
-  s: () => ({ strikethrough: true }),
-  strong: () => ({ bold: true }),
-  u: () => ({ underline: true }),
-}
+import { config as defaultConfig, Config } from '../../config/htmlToSlate/default'
 
 const deserialize = (
   el: ChildNode,
-  {
-    attributeMap = [],
-  }: {
-    attributeMap?: IattributeMap[]
-  } = {},
+  config: Config = defaultConfig,
 ): any => {
   if (el.type !== ElementType.Tag && el.type !== ElementType.Text) {
     return null
@@ -57,16 +21,16 @@ const deserialize = (
 
   const nodeName = getName(parent)
 
-  const children = parent.childNodes ? parent.childNodes.map((node) => deserialize(node, { attributeMap })).flat() : []
+  const children = parent.childNodes ? parent.childNodes.map((node) => deserialize(node, config)).flat() : []
 
   if (getName(parent) === 'body') {
     return jsx('fragment', {}, children)
   }
 
-  if (ELEMENT_TAGS[nodeName]) {
-    let attrs = ELEMENT_TAGS[nodeName](parent)
+  if (config.elementTags[nodeName]) {
+    let attrs = config.elementTags[nodeName](parent)
     // tslint:disable-next-line no-unused-expression
-    attributeMap.map((map) => {
+    /*attributeMap.map((map) => {
       const value = getAttributeValue(parent, map.htmlAttr)
       if (value) {
         attrs = {
@@ -74,11 +38,11 @@ const deserialize = (
           ...attrs,
         }
       }
-    })
+    })*/
     return jsx('element', attrs, children)
   }
 
-  if (TEXT_TAGS[nodeName] || el.type === ElementType.Text) {
+  if (config.textTags[nodeName] || el.type === ElementType.Text) {
     const text = textContent(el)
     if (!hasLineBreak(text) && text.trim() === '') {
       return null
@@ -89,13 +53,16 @@ const deserialize = (
   return children
 }
 
-const gatherTextMarkAttributes = (el: Element) => {
+const gatherTextMarkAttributes = (
+  el: Element,
+  config: Config = defaultConfig
+  ) => {
   let allAttrs = {}
   // tslint:disable-next-line no-unused-expression
   if (el.childNodes) {
     ;[el, ...getChildren(el).flat()].forEach((child) => {
       const name = getName(child as Element)
-      const attrs = TEXT_TAGS[name] ? TEXT_TAGS[name](child as Element) : {}
+      const attrs = config.textTags[name] ? config.textTags[name](child as Element) : {}
       const text = textContent(child)
       allAttrs = {
         ...allAttrs,
@@ -105,7 +72,7 @@ const gatherTextMarkAttributes = (el: Element) => {
     })
   } else {
     const name = getName(el)
-    const attrs = TEXT_TAGS[name] ? TEXT_TAGS[name](el) : {}
+    const attrs = config.textTags[name] ? config.textTags[name](el) : {}
     const text = textContent(el)
     allAttrs = {
       ...attrs,
@@ -117,11 +84,7 @@ const gatherTextMarkAttributes = (el: Element) => {
 
 export const htmlToSlate = (
   html: string,
-  {
-    attributeMap = [],
-  }: {
-    attributeMap?: IattributeMap[]
-  } = {},
+  config: Config = defaultConfig,
 ) => {
   let slateContent
   const handler = new DomHandler((error, dom) => {
@@ -130,7 +93,7 @@ export const htmlToSlate = (
     } else {
       // Parsing completed, do something
       slateContent = dom
-        .map((node) => deserialize(node, { attributeMap })) // run the deserializer
+        .map((node) => deserialize(node, config)) // run the deserializer
         .filter((element) => element) // filter out null elements
         .map((element) => {
           // ensure all top level elements have a children property
